REFACTOR: abstract Resolving::Emission for packs and provisions
