+++ lib/spaces/engines/requires.rb
@@ -4,13 +4,15 @@ require_level 'spaces/engines/transforming/divisions/bindings'
 require_all   'spaces/engines/transforming/emissions'
 
 require_all   'spaces/engines/arenas'
-require_all   'spaces/engines/packing'
+require       'spaces/engines/resolving/emission' #FIX!: kludge!
+require_all   'spaces/engines/packing' #FIX!: this is too early in the layers
 require_all   'spaces/engines/publishing'
 require_all   'spaces/engines/blueprinting'
 require_all   'spaces/engines/locating'
 require_all   'spaces/engines/settling'
 require_all   'spaces/engines/installing'
 require_all   'spaces/engines/resolving'
+# require_all   'spaces/engines/packing' #FIX!: this should be here
 require_all   'spaces/engines/provisioning'
 
 require_all   'spaces/engines/provider_aspects'
